pyproject.toml
[project]
name = "inukai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.10"
dependencies = [
    "amazon-textract-textractor>=1.8.5",
    "boto3-stubs>=1.35.81",
    "boto3>=1.35.78",
    "click>=8.1.7",
    "editables>=0.5",
    "geopy>=2.4.1",
    "hatchling>=1.26.3",
    "llava",
    "loguru>=0.7.3",
    "opencage>=3.0.3",
    "pandas>=2.2.3",
    "protobuf>=5.29.1",
    "setuptools>=75.6.0",
    "thefuzz>=0.22.1",
    "torch==2.1.2",
    "transformers==4.37.2",
    "mypy-boto3-textract>=1.35.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
validate = "inukai.validate.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
llava = { git = "https://github.com/haotian-liu/LLaVA" }
types-boto3-custom = { path = "vendored/types_boto3_custom-1.35.82-py3-none-any.whl" }

[dependency-groups]
dev = [
    "basedpyright>=1.22.1",
    "types-boto3-custom",
]
